Self assignment of local variable detected JAVA-E0291
A local variable is assigned to itself.

This is essentially a noop but it may be indicative of a different problem. It may be that the variable shadows another in a parent scope, or that the variable may shadow a field of the object itself. Such code can cause confusion and subtle logic errors that are hard to catch.