Getter or setter methods must not perform operations other than reading or writing fields they are named after.
The main purpose of a getter is to allow easy access to private fields that the programmer wants to expose. Setters are, similarly, meant to allow users of a class to set the values of private fields.

When working with multiple fields that all need getters and/or setters, it is easy to make the mistake of just copy-pasting the same method, and forgetting to change the field that is returned/assigned, or forgetting to change the name of the method. Such mistakes can cost a great amount of time to find.

This issue is raised when:

A getter does not read the field it is named after.
A setter does not write to the field it is named after.

The function getImu did not return an IMU object. It has been renamed to getImuAngle